Created Camara class to use camara api
