+++ app/src/main/java/com/example/easycloset/Camera.java
@@ -1,0 +1,53 @@
+package com.example.easycloset;
+
+import android.content.Context;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Environment;
+import android.provider.MediaStore;
+
+import androidx.core.content.FileProvider;
+
+import java.io.File;
+
+public class Camera {
+
+    Context context;
+    private File photoFile;
+    Constant constant;
+    public static String TAG = ".Camera";
+
+    public Camera(Context context, File photoFile) {
+        this.context = context;
+        this.photoFile = photoFile;
+    }
+
+    private void launchCamera() {
+        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+        // Create a File reference for future access
+        String photoFileName = "photo.jpg";
+        photoFile = getPhotoFileUri(photoFileName);
+        // wrap File object into a content provider
+        Uri fileProvider = FileProvider.getUriForFile(context, "com.codepath.fileprovider", photoFile);
+        intent.putExtra(MediaStore.EXTRA_OUTPUT, fileProvider);
+        // If you call startActivityForResult() using an intent that no app can handle, your app will crash.
+        // So as long as the result is not null, it's safe to use the intent.
+        if (intent.resolveActivity(context.getPackageManager()) != null) {
+            // Start the image capture intent to take photo
+            ((MainActivity) context).startActivityForResult(intent, constant.getCAPTURE_IMAGE_ACTIVITY_REQUEST_CODE());
+        }
+    }
+
+    private File getPhotoFileUri(String photoFileName) {
+        // Get safe storage directory for photos
+        // Use `getExternalFilesDir` on Context to access package-specific directories.
+        // This way, we don't need to request external read/write runtime permissions.
+        File mediaStorageDir = new File(context.getExternalFilesDir(Environment.DIRECTORY_PICTURES), TAG);
+        // Create the storage directory if it does not exist
+        if (!mediaStorageDir.exists()) {
+            mediaStorageDir.mkdirs();
+        }
+        // Return the file target for the photo based on filename
+        return new File(mediaStorageDir.getPath() + File.separator + photoFileName);
+    }
+}
