+++ app/src/main/java/com/example/easycloset/Constant.java
@@ -3,8 +3,13 @@ package com.example.easycloset;
 public class Constant {
 
     private final int DELETE_ITEM_REQUEST_CODE = 7;
+    private  final int CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE = 20;
 
     public int getDELETE_ITEM_REQUEST_CODE() {
         return DELETE_ITEM_REQUEST_CODE;
+    }
+
+    public int getCAPTURE_IMAGE_ACTIVITY_REQUEST_CODE() {
+        return CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE;
     }
 }
