+++ app/src/main/java/com/example/easycloset/UploadFragment.java
@@ -94,21 +94,6 @@ public class UploadFragment extends Fragment {
         });
     }
 
-    private void launchCamera() {
-        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        // Create a File reference for future access
-        String photoFileName = "photo.jpg";
-        photoFile = getPhotoFileUri(photoFileName);
-        // wrap File object into a content provider
-        Uri fileProvider = FileProvider.getUriForFile(requireContext(), "com.codepath.fileprovider", photoFile);
-        intent.putExtra(MediaStore.EXTRA_OUTPUT, fileProvider);
-        // If you call startActivityForResult() using an intent that no app can handle, your app will crash.
-        // So as long as the result is not null, it's safe to use the intent.
-        if (intent.resolveActivity(requireContext().getPackageManager()) != null) {
-            // Start the image capture intent to take photo
-            startActivityForResult(intent, CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE);
-        }
-    }
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
@@ -126,18 +111,7 @@ public class UploadFragment extends Fragment {
         }
     }
 
-    private File getPhotoFileUri(String photoFileName) {
-        // Get safe storage directory for photos
-        // Use `getExternalFilesDir` on Context to access package-specific directories.
-        // This way, we don't need to request external read/write runtime permissions.
-        File mediaStorageDir = new File(requireContext().getExternalFilesDir(Environment.DIRECTORY_PICTURES), TAG);
-        // Create the storage directory if it does not exist
-        if (!mediaStorageDir.exists()) {
-            mediaStorageDir.mkdirs();
-        }
-        // Return the file target for the photo based on filename
-        return new File(mediaStorageDir.getPath() + File.separator + photoFileName);
-    }
+
 
     @SuppressLint("NotifyDataSetChanged")
     private void saveItem(String colour, String category, File photoFile) {
